feat: adicionar validação e formatação para CPF, CNPJ e telefone, incluindo documentação e melhorias na estrutura
