+++ frappe_agt/public/ts/utils/brazil/index.ts
@@ -2,3 +2,4 @@ import "./cnpj";
 import "./cpf";
 import "./cpf_cnpj";
 import "./cep";
+import "./phone";
